+++ src/ComplexSearch/ServiceSearch.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { cloneElement } from 'react'
 import PropTypes from 'prop-types'
 import classnames from 'classnames'
 import { injectSheet } from '../theme'
@@ -65,7 +65,6 @@ import provideSearchDropdown from './provideSearchDropdown'
     '&::-ms-reveal, &::-ms-clear': {
       display: 'none'
     },
-
     '&::-webkit-input-placeholder': {
       fontSize: theme.serviceSearch.input.placeholder.fontSize,
       color: theme.serviceSearch.input.placeholder.color,
@@ -82,14 +81,28 @@ import provideSearchDropdown from './provideSearchDropdown'
       opacity: 1
     }
   },
-  searchIcon: {
-    color: theme.serviceSearch.input.default.icon,
+  inputLeftIcon: {
+    marginLeft:  12
+  },
+  searchButton: {
+    extend: isolateMixin,
+    background: 'none',
+    outline: 'none',
+    border: 'none',
+    cursor: 'pointer',
     position: 'absolute',
+    padding: 0,
+    width: 15,
+    height: 15,
     right: 15,
     top: '50%',
     transform: 'translateY(-50%)'
   },
-  сlearIcon: {
+  searchIcon: {
+    color: theme.serviceSearch.input.default.icon,
+    outline: 'none'
+  },
+  clearIcon: {
     position: 'absolute',
     right: 15,
     top: '50%',
@@ -201,7 +214,11 @@ export default class ServiceSearch extends React.Component {
     /**
      * Размер поискового блока
      */
-    size: PropTypes.oneOf(['small', 'medium'])
+    size: PropTypes.oneOf(['small', 'medium']),
+    /**
+    * Иконка инпута слева
+     */
+    inputLeftIcon: PropTypes.node
   };
 
   static defaultProps = {
@@ -238,6 +255,18 @@ export default class ServiceSearch extends React.Component {
     this.setState({sourceType: type})
   }
 
+  renderInputIcon() {
+    const {inputLeftIcon, theme, classes} = this.props
+    if (!inputLeftIcon)
+      return
+    const {size, className, color} = inputLeftIcon.props
+    return cloneElement(inputLeftIcon, {
+      className: classnames(classes.inputLeftIcon, className),
+      size: size || 15,
+      color: color || theme.search.input.default.icon
+    })
+  }
+
   renderInputNode() {
     const {
       placeholder,
@@ -267,6 +296,30 @@ export default class ServiceSearch extends React.Component {
     )
   }
 
+  renderButton() {
+    const {
+      classes,
+      searchButtonClassName,
+      searchButtonProps,
+      onSubmit
+    } = this.props
+
+    return (
+      <button
+        className={classnames(classes.searchButton, searchButtonClassName)}
+        onClick={onSubmit}
+        {...searchButtonProps}
+      >
+        <ServiceSearchIcon
+          size={15}
+          className={classes.searchIcon}
+          tabIndex={-1}
+          color="currentColor"
+        />
+      </button>
+    )
+  }
+
   renderInput = () => {
     const {
       inputWrapperClassName,
@@ -280,15 +333,12 @@ export default class ServiceSearch extends React.Component {
         inputWrapperClassName,
         isDropdownOpened && classes.active
       )}>
+        {this.renderInputIcon()}
         {this.renderInputNode()}
-        {!this.isClearVisible && <ServiceSearchIcon
-          size="15"
-          className={classes.searchIcon}
-          color="currentColor"
-        />}
+        {!this.isClearVisible && this.renderButton()}
         {this.isClearVisible && <ClearIcon
-          className={classes.сlearIcon}
-          size="20"
+          className={classes.clearIcon}
+          size={20}
           color="currentColor"
           onClick={this.props.clearForm}
         ></ClearIcon>}