+++ src/ComplexSearch/ServiceSearch.test.js
@@ -67,10 +67,4 @@ describe('<ServiceSearch />', () => {
     input.simulate('keydown', {key: 'Enter'})
     expect(handlersProps.onPressEnter).toHaveBeenCalled()
   })
-
-  it('should not render button', () => {
-    const wrapper = getSearchWrapper()
-    expect(wrapper.find('button').length).toEqual(0)
-  })
-
 })