+++ src/Select/Select.js
@@ -48,11 +48,11 @@ const multipleSelectFix = <optgroup disabled hidden />
         pointerEvents: 'none'
       }
     },
-    ...placeholderMixin('&$isReadonly:not($isDisabled) $input input', {
+    ...placeholderMixin('&$isReadonly:not($lightPlaceholder):not($isDisabled) $input input', {
       opacity: 1,
       color: theme.field.colors.default.text
     }),
-    ...placeholderMixin('&:not($isFocused):not($isDisabled) $input input', {
+    ...placeholderMixin('&:not($isFocused):not($lightPlaceholder):not($isDisabled) $input input', {
       opacity: 1,
       color: theme.field.colors.default.text
     })
@@ -236,7 +236,8 @@ const multipleSelectFix = <optgroup disabled hidden />
   isMultipleDropdown: {},
   withSearch: {},
   withLeftIcon: {},
-  withRightIcon: {}
+  withRightIcon: {},
+  lightPlaceholder: {}
 }), {name: 'Select'})
 export default class Select extends PureComponent {
 
@@ -299,6 +300,10 @@ export default class Select extends PureComponent {
      * Плэйсхолдер
      */
     placeholder: PropTypes.string,
+    /**
+     * Светлый плейсхолдер, как в компоненте `<Input />`
+     */
+    lightPlaceholderColor: PropTypes.bool,
     /**
      * Доступность элемента
      */
@@ -627,6 +632,7 @@ export default class Select extends PureComponent {
       native,
       clearIcon,
       inputMode,
+      lightPlaceholderColor,
       /* eslint-enable no-unused-vars */
       ...props
     } = this.props
@@ -785,7 +791,8 @@ export default class Select extends PureComponent {
       disabled,
       size,
       icon,
-      classes
+      classes,
+      lightPlaceholderColor
     } = this.props
 
     const focusedInput = inputFocused || isOpened
@@ -804,6 +811,7 @@ export default class Select extends PureComponent {
       disabled && classes.isDisabled,
       isOpened && classes.isOpened,
       focusedInput && classes.isFocused,
+      lightPlaceholderColor && classes.lightPlaceholder,
       multiple && !onSearch && classes.isMultipleWithoutSearch
     )
 
@@ -917,7 +925,8 @@ export default class Select extends PureComponent {
       className,
       classes,
       rootStyle,
-      rootClassName
+      rootClassName,
+      lightPlaceholderColor
     } = this.props
 
     const {
@@ -956,7 +965,8 @@ export default class Select extends PureComponent {
       disabled && classes.isDisabled,
       inputFocused && classes.isFocused,
       multiple && classes.isMultipleWithoutSearch,
-      multiple && classes.withCustom
+      multiple && classes.withCustom,
+      lightPlaceholderColor && classes.lightPlaceholder
     )
 
     return (